requirements.txt
lxml>=5.3.0
pydantic>=2.10.0
click>=8.1.0
rich>=13.0.0
markdown>=3.5.0
PyYAML>=6.0.0
